// src/app/interfaces/models.ts
export class UserModel {
    id: number;
    nom: string;
    prenom: string;
    login: string;
    password?: string;
    sex: string;
    date_naissance: string;
    photoUser: string;
    typeCompte: string;
    date_created: string;
    date_update: string;
    token?: string;
    pseudo_name: string;
  }